+++ prisma/migrations/20260510205743_add_vehicle_customer_relation/migration.sql
@@ -1,0 +1,6 @@
+-- AlterTable
+ALTER TABLE "vehicles" ADD COLUMN     "customerId" TEXT,
+ADD COLUMN     "notes" TEXT;
+
+-- AddForeignKey
+ALTER TABLE "vehicles" ADD CONSTRAINT "vehicles_customerId_fkey" FOREIGN KEY ("customerId") REFERENCES "customers"("id") ON DELETE SET NULL ON UPDATE CASCADE;